+++ ydb/core/tx/columnshard/engines/reader/simple_reader/duplicates/ya.make
@@ -8,6 +8,7 @@ SRCS(
     executor.cpp
     filters.cpp
     manager.cpp
+    merge.cpp
     private_events.cpp
 )
